Fix formation saves rebuilding defense
